requirements.txt
numpy>=1.25
pandas>=2.1
pyyaml>=6.0
scikit-learn>=1.3
torch>=2.1
matplotlib>=3.8
tqdm>=4.66
